move specific l2 func to payload builder
